+++ deeptab/architectures/resnet.py
@@ -72,7 +72,7 @@ class ResNet(BaseModel):
                 *feature_information,
                 config=config,
             )
-            input_dim = np.sum([len(info) * self.hparams.d_model for info in feature_information])
+            input_dim = int(np.sum([len(info) * self.hparams.d_model for info in feature_information]))
         else:
             input_dim = get_feature_dimensions(*feature_information)
