+++ deeptab/architectures/tabm.py
@@ -79,7 +79,7 @@ class TabM(BaseModel):
             if self.hparams.average_embeddings:
                 input_dim = self.hparams.d_model
             else:
-                input_dim = np.sum([len(info) * self.hparams.d_model for info in feature_information])
+                input_dim = int(np.sum([len(info) * self.hparams.d_model for info in feature_information]))
 
         else:
             input_dim = get_feature_dimensions(*feature_information)
